+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### 新特性
 * [core] 优化BeanDesc，适配更多Getter和Setter方法
+* [extra] 增加基于zxing的二维码生成和解码（zxing可选依赖）
 
 ### Bug修复
 * [setting] 修复clear方法未清空group的问题，store方法未换行问题，set方法分组丢失问题（感谢@【广西】Succy）